+++ client/src/components/common/PricePannel/index.tsx
@@ -2,9 +2,6 @@ import React, { useState, useContext } from 'react';
 import useLocation from '@/hooks/customHooks/useLocation';
 import styled from '@emotion/styled';
 
-import { FormContext } from '@/hooks/context';
-
-import OrderApi from '@/apis/OrderApi';
 import { baeminFont, normalRadius, primary1, white } from '@/static/style/common';
 import useHistory from '@/hooks/customHooks/useHistory';
 
@@ -13,8 +10,6 @@ type PricePannelProps = {
 };
 
 const PricePannel = ({ productTotalPrice }: PricePannelProps) => {
-  const { form } = useContext(FormContext) || {};
-
   const curLocation = useLocation();
   const history = useHistory();
   const deliveryCost = productTotalPrice < 30000 ? 2500 : 0;
@@ -22,10 +17,6 @@ const PricePannel = ({ productTotalPrice }: PricePannelProps) => {
   const handleClickOrderBtn = async () => {
     if (curLocation === '/cart') {
       history.push('/order');
-    } else if (curLocation === '/order') {
-      const paresedData = JSON.parse(localStorage.getItem('cartInfo'));
-      const { cartId, product } = paresedData;
-      await OrderApi.order({ id: cartId, ...form, selectedItem: Object.values(product) });
     }
   };
