+++ client/src/pages/Order/index.tsx
@@ -1,50 +1,17 @@
-import React, { createContext } from 'react';
+import React from 'react';
 import styled from '@emotion/styled';
 
-import { FormContext } from '@/hooks/context';
-import useInput from '@/hooks/customHooks/useInput';
-import Validation from '@/utils/validation';
-
-import PricePannel from '@/components/common/PricePannel';
 import OrderStageHeader from '@/components/common/OrderStageHeader';
 import { OrderForm } from '@/components/Order';
 import OrderSummary from '@/components/Order/OrderSummary';
 
-const stage1InitialForm = {
-  buyerName: '',
-  phone: '',
-  email: '',
-};
-
-const stage2InitialForm = {
-  receiverName: '',
-  receiverAddress: '',
-  receiverPhone: '',
-};
-
-const validationSchema = {
-  buyerName: Validation().require().isName(),
-  phone: Validation().require().isPhone(),
-  email: Validation().require().isEmail(),
-  receiverName: Validation().require().isName(),
-  receiverAddress: Validation().require(),
-  receiverPhone: Validation().require().isPhone(),
-};
-
 const Order = () => {
-  const { form, onChange, onBlur, check, error } = useInput({
-    initialState: { ...stage1InitialForm, ...stage2InitialForm },
-    validationSchema,
-  });
-
   return (
     <OrderContainer>
       <OrderStageHeader title="주문하기" />
       <OrderBody>
-        <FormContext.Provider value={{ form, onChange, onBlur, check, error }}>
-          <OrderForm />
-          <OrderSummary />
-        </FormContext.Provider>
+        <OrderForm />
+        <OrderSummary />
       </OrderBody>
     </OrderContainer>
   );
